EYPrototype/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6

langchain==0.3.0
langchain-openai==0.2.0
openai==1.50.0

pydantic==2.9.0
python-dotenv==1.0.0
requests==2.31.0

pymongo==4.6.1
